gait_render_auto_yolo/requirements.txt
flask
pillow
gunicorn
ultralytics
